# Conflicts:
#	app/Http/Controllers/V2/InstanceController.php
#	docs/v2/public-openapi.yaml

+++ app/Http/Controllers/V2/InstanceController.php
@@ -7,7 +7,11 @@ use App\Http\Requests\V2\Instance\CreateRequest;
 use App\Http\Requests\V2\Instance\DeployRequest;
 use App\Http\Requests\V2\Instance\UpdateRequest;
 use App\Models\V2\Instance;
+use App\Resources\V2\CredentialResource;
 use App\Resources\V2\InstanceResource;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use UKFast\DB\Ditto\QueryTransformer;
 
 /**
@@ -52,7 +56,10 @@ class InstanceController extends BaseController
      */
     public function store(CreateRequest $request)
     {
-        $instance = new Instance($request->only(['network_id', 'name']));
+        $instance = new Instance($request->only(['name', 'vpc_id', 'locked']));
+        if (!$request->has('locked')) {
+            $instance->locked = false;
+        }
         $instance->save();
         $instance->refresh();
         return $this->responseIdMeta($request, $instance->getKey(), 201);
@@ -66,7 +73,12 @@ class InstanceController extends BaseController
     public function update(UpdateRequest $request, string $instanceId)
     {
         $instance = Instance::forUser(app('request')->user)->findOrFail($instanceId);
-        $instance->fill($request->only(['vpc_id', 'name']));
+        if (!$this->isAdmin &&
+            (!$request->has('locked') || $request->get('locked') !== false) &&
+            $instance->locked === true) {
+            return $this->isLocked();
+        }
+        $instance->fill($request->only(['name', 'vpc_id', 'locked']));
         $instance->save();
         return $this->responseIdMeta($request, $instance->getKey(), 200);
     }
@@ -79,8 +91,41 @@ class InstanceController extends BaseController
     public function destroy(Request $request, string $instanceId)
     {
         $instance = Instance::forUser($request->user)->findOrFail($instanceId);
+        if (!$this->isAdmin && $instance->locked === true) {
+            return $this->isLocked();
+        }
         $instance->delete();
         return response()->json([], 204);
+    }
+
+    /**
+     * @return \Illuminate\Http\JsonResponse
+     */
+    private function isLocked(): JsonResponse
+    {
+        return JsonResponse::create([
+            'errors' => [
+                'title'  => 'Forbidden',
+                'detail' => 'The specified instance is locked',
+                'status' => 403,
+            ]
+        ], 403);
+    }
+
+    /**
+     * @param  Request  $request
+     * @param  string  $instanceId
+     *
+     * @return AnonymousResourceCollection|\Illuminate\Support\HigherOrderTapProxy|mixed
+     */
+    public function credentials(Request $request, string $instanceId)
+    {
+        return CredentialResource::collection(
+            Instance::forUser($request->user)
+                ->findOrFail($instanceId)
+                ->credentials()
+                ->paginate($request->input('per_page', env('PAGINATION_LIMIT')))
+        );
     }
 
     public function deploy(DeployRequest $request, string $instanceId)
